requirements.txt
django
djangorestframework
sentence-transformers
faiss-cpu
openai
numpy
google-generativeai
pypdf
python-docx
djangorestframework-simplejwt
requests
dotenv
beautifulsoup4